Add requests-mock fixtures for offline tests
