+++ pyproject.toml
@@ -37,6 +37,7 @@ ruff = "^0.1.3"
 pdoc = "^15.0.0"
 pytest = "^7.4.3"
 pyright = "^1.1.364"
+requests-mock = "^1.11.0"
 
 #[tool.poetry.extras]
 #df = ["pandas"]
