+++ tests/fixtures/sample_response.json
@@ -1,0 +1,8 @@
+{
+  "totalCount": 1000,
+  "data": [
+    {"s": "SYM0", "d": ["Name0", 100, 1000, 0.0]},
+    {"s": "SYM1", "d": ["Name1", 101, 1001, 1.0]},
+    {"s": "SYM2", "d": ["Name2", 102, 1002, 2.0]}
+  ]
+}
